Moved thread related classes to root
