+++ src/main/java/univalle/tedesoft/uno/controller/GameController.java
@@ -16,7 +16,7 @@ import univalle.tedesoft.uno.model.Players.MachinePlayer;
 import univalle.tedesoft.uno.model.Players.Player;
 import univalle.tedesoft.uno.model.State.GameState;
 import univalle.tedesoft.uno.model.State.IGameState;
-import univalle.tedesoft.uno.model.Threads.MachinePlayerRunnable;
+import univalle.tedesoft.uno.threads.MachinePlayerRunnable;
 import univalle.tedesoft.uno.view.GameView;
 
 import java.util.List;
